requirements.txt
filetype==1.2.0
numpy==1.24.2
Pillow==9.4.0
scipy==1.10.1
